requirements.txt
django
requests
python-decouple
python-dotenv
djangorestframework